refactor(frontend): spotifyAuthの中身をspotifyServiceに移動 #187
